Convert dynamic network's capacity to longs. FluidNetwork still clamped to integer for now though. Also fix some potentially incorrect calculations for energy network scale at large network capacities

+++ src/api/java/mekanism/api/transmitters/IGridTransmitter.java
@@ -43,12 +43,12 @@ public interface IGridTransmitter<ACCEPTOR, NETWORK extends DynamicNetwork<ACCEP
 
     ITextComponent getTransmitterNetworkBuffer();
 
-    int getTransmitterNetworkCapacity();
+    long getTransmitterNetworkCapacity();
 
     @Nonnull
     FloatingLong getCapacityAsFloatingLong();
 
-    int getCapacity();
+    long getCapacity();
 
     World world();
